backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
faiss-cpu==1.8.0.post1
scikit-learn==1.5.2
httpx==0.27.2
requests==2.32.3
pyyaml==6.0.2
# Local AI
torch==2.4.1
sentence-transformers==3.0.1
transformers==4.44.2
# Optional: for CSV/Parquet speedups
pyarrow==17.0.0
fastparquet==2024.5.0
# PDF ingestion
pypdf==4.3.1
pdfminer.six==20231228
# HTML parsing for downloader (kept if ever needed)
beautifulsoup4==4.12.3
